Debugger updates

+++ ICSharpCode.Decompiler/CodeMappings.cs
@@ -54,7 +54,7 @@ namespace ICSharpCode.Decompiler
 		/// Retrieves the array that contains the IL range and the missing gaps between ranges.
 		/// </summary>
 		/// <returns>The array representation of the step aranges.</returns>
-		public int[] ToArray(bool isMatch)
+		public uint[] ToArray(bool isMatch)
 		{
 			var currentList = new List<ILRange>();
 			
@@ -130,7 +130,7 @@ namespace ICSharpCode.Decompiler
 			}
 		}
 
-		public int[] ToArray(List<ILRange> currentList, bool isMatch)
+		public uint[] ToArray(List<ILRange> currentList, bool isMatch)
 		{
 			if (currentList == null)
 				currentList = new List<ILRange>();
@@ -146,10 +146,10 @@ namespace ICSharpCode.Decompiler
 			}
 			
 			// set the output
-			var resultList = new List<int>();
+			var resultList = new List<uint>();
 			foreach (var element in ILRange.OrderAndJoin(currentList)) {
-				resultList.Add((int)element.From);
-				resultList.Add((int)element.To);
+				resultList.Add(element.From);
+				resultList.Add(element.To);
 			}
 			
 			return resultList.ToArray();
